requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
Bottleneck==1.6.0
burnysc2==7.1.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
frozenlist==1.8.0
graphviz==0.21
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.1
loguru==0.7.3
MarkupSafe==3.0.3
matplotlib==3.10.7
mpyq==0.2.5
multidict==6.7.0
numba==0.62.1
numexpr==2.14.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
portpicker==1.6.0
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
s2clientprotocol==5.0.15.95299.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
tabulate==0.9.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
win32_setctime==1.2.0
yarl==1.22.0
pyarrow~=22.0.0